of: reserved_mem: Move MAX_RESERVED_REGIONS to header file
MAX_RESERVED_REGIONS is a useful constant for clients of OF reserved
memory regions that provide their own initialisation function callback
using RESERVEDMEM_OF_DECLARE(), since it allows them to avoid dynamic
allocation of book-keeping structures.

Move the constant to linux/of_reserved_mem.h and tweak the name to
"OF_MAX_RESERVED_MEM_REGIONS" so that it's less likely to conflict with
an unrelated constant in the kernel.

Cc: Rob Herring <robh+dt@kernel.org>
Cc: Frank Rowand <frowand.list@gmail.com>
Signed-off-by: Will Deacon <will@kernel.org>