kvm: arm64: Add user ABI to attach to SPCI partition
Using some bits of the `type` argument to KVM_CREATE_VM, user space can
express its intent to attach to a partition in order to schedule the
vCPUs and further interact with it.
---
There needs to be some means of access control on this. E.g. in Andorid,
we'll want SELinux rules so only the designated process can attached to
any give partition. I don't know how do do that so I haven't done
anything about it yet.