arm64: kvm: Introduce run loop for the host
With the init vector installed, the call to __do_hyp_init no longer
simply returns to the host directly with an eret but instead begins to
treat the host as a vCPU and repeatedly __guest_enters into it.

As a result, hyp is endowed with its very own context for the general
purpose registers. However, this change stores the state in a confusing
way:
   - host gp_regs are stored in the kvm_host_vcpu
   - hyp gp_regs are store in the kvm_host_data context
   - other host registers are store in the kvm_host_data context

Doing it this way allows the initial change to be less intrusive
although this will be neated up so the host state only resides within
kvm_host_vcpu.

Signed-off-by: Andrew Scull <ascull@google.com>