arm64: kvm: Migrate hyp interface to SMCCC
Rather than passing arbitrary function pointers to run at hyp, define
and equivalent set of SMCCC functions. Since the SMCCC functions are
strongly tied to true function prototypes, it is not expected for the
host to ever call an invalid ID but a warning is generared if this does
ever occur.

Signed-off-by: Andrew Scull <ascull@google.com>
Signed-off-by: David Brazdil <dbrazdil@google.com>