arm64: kvm: Move __create_hyp_mappings() to header file
In preparation for being able to re-use the stage 1 hyp page table code
at EL2, move all the necessary helpers to a standalone header file.
This header is meant to be used as a template, and enables the future
users to intercept some of the symbols, notably for page allocation.

Signed-off-by: Quentin Perret <qperret@google.com>