arm64: kvm: Dynamically allocate kvm_arm_hyp_percpu_base
The addresses of host-allocated pages for the hyp per-cpu data need to
be passed at EL2 as they will need to be re-mapped when re-creating the
new page tables.

To simplify this procedure, change kvm_arm_hyp_percpu_base into a
dynamically allocated (and hence contiguous) array holding the PA of the
allocated pages. While a it, map the array itself in the initial
pgtables to make it accessible from the hyp init code.

Signed-off-by: Quentin Perret <qperret@google.com>