kvm: arm64: Extract parts of el2_setup into hyp_setup
When the a CPU is booted in EL2, the kernel checks for VHE support and
initializes the CPU core accordingly. For nVHE it also installs the stub
vectors and drops down to EL1.

Once KVM gains the ability to boot cores without going through the
kernel entry point, it will need to initialize the CPU the same way.
Extract the relevant bits of el2_setup into hyp_setup inside the
kvm/hyp/.

The hyp build rules compile the function for VHE and nVHE separately
with __KVM_VHE/NVHE_HYPERVISOR__ defined. Use this to specialize each
copy of the function.

Signed-off-by: David Brazdil <dbrazdil@google.com>