KVM: arm64: Function to verify memory is host's
Add a function to check that the specified memory is host
memory and not a protected VM's, device, or invalid.

EL2 cannot trust that host provided memory addresses and
pointers are legitimate. Future patches need this to check that
pointers to host VM state are safe to access.

Currently this check is very basic. Will think it through a bit
more.

Signed-off-by: Fuad Tabba <tabba@google.com>