KVM: arm64: Shadow table for KVM EL2 state
Create and populate a shadow table for struct kvm and struct
kvm_vcpu at EL2. The memory for the structs is donated by the
host and then unmapped from the host at stage 1 and at stage 2.

Currently the EL2 state isn't used. It's only sanity-checked
against the state from the host.

Signed-off-by: Fuad Tabba <tabba@google.com>