ANDROID: KVM: arm64: Shadow table for KVM EL2 state
Create and populate a shadow table that contains the state hyp
need for running protected VMs, i.e., struct kvm and struct
kvm_vcpu at EL2.

The memory for this is donated by the host and then unmapped from
the host at stage 1 and at stage 2 (by hyp).

This state is not used yet.

Signed-off-by: Fuad Tabba <tabba@google.com>
Signed-off-by: Quentin Perret <qperret@google.com>
Bug: 209580772
Change-Id: I973d2516e6ecd0da315e116236f54204ee9b5c0c