arm64: mm: add 'address' parameter to page table manipulation helpers
In order to permit the hypervisor to reason about whether certain page
table updates should be permitted, we need to provide the target virtual
address when available. Unfortunately, in some cases, we cannot access
the address as doing so would involve making changes to core code, but
we can deal with this shortcoming at the hyp level.

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>