KVM: arm64: refactor: create a core state for vcpus
The kvm_vcpu_arch contains various kinds of data relevant to the
guest VM, some of it has to do with the core state of the virtual
machine, such as its context, some of it is relevant to the host
to schedule and manage the guest's communication with the outside
world, and some of it is for debug, trace, and statistics.

Factor out the core state relevant data into a substructure. This
facilitates using only that structure in future patches in this
series to protect only that state for protected VMs.

Refactoring only. No functional changes intended.

Signed-off-by: Fuad Tabba <tabba@google.com>