fs: kconfig: relax hard dependency on generic crypto algorithms
The kconfig dependencies for fsverity and fs encryption are currently
stricter than necessary: the crypto API may provide different
implementations of the same algorithm, and so using 'depends on' for the
generic versions is overkill, and we can use imply to relax it to what
amounts to a recommendation. On architectures such as arm64, all the
algorithms in question have native implementation that are faster and
more efficient.

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>