arm64: module: preserve RELA sections FIPS140 module integrity selfcheck
The FIPS 140-2 integrity check compares the runtime code with a digest
that was created at build time. Given that the module's placement in
virtual memory is an a priori unknown, we cannot account for this at
build time, and so we need to do so at runtime instead.

In order to revert the code to the build time state, we need to know
which changes the module loader applied to it. These changes are based
on the RELA ELF section that describes the changes that the module
loader must apply, and so to unapply these changes, we need to preserve
the RELA section when loading the module.

So add a special case for a module called 'fips140' in the module
loader, and copy the RELA section applying to .text to a temporary
buffer that the fips140.ko init code can access.

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>