crypto: drbg - avoid subsys_initcall for generic DRBG driver
The generic DRBG driver is not used by the fuzz tests, given that their
output is not reproducible, and so comparing the generic implementation
with an architecture specific one based on the output is not possible.

This means it is not necessary to initialize the DRBG module earlier
than usual when it is builtin.

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>