arm64: alternatives: add a flavor that always emits the default sequence
The ALTERNATIVE() macro takes an optional CONFIG_ option name that when
specified, will prevent the entire sequence from being emitted if its
value evaluates to false. This is useful for preventing strings of NOPs
from being emitted into the binary for features that are not enabled in
the build to begin with.

In some cases, we would like the default sequence to be emitted
unconditionally, and only include the alternative sequence and the
patching metadata if a certain feature is enabled in the build. Let's
add support for that in a new ALTERNATIVE_IF() macro.

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>