KVM: arm64: create and use a new vcpu_hyp_state struct
Create a struct for the hypervisor state from the related fields
in vcpu_arch. This is needed in future patches to reduce the
scope of functions from the vcpu as a whole to only the relevant
state, via this newly created struct.

Create a new instance of this struct in vcpu_arch and fix the
accessors to use the new fields. Remove the existing fields from
vcpu_arch.

Signed-off-by: Fuad Tabba <tabba@google.com>