KVM: arm64: Handle FFA_RXTX_MAP.
This maps the FF-A RX and TX buffers into the hypervisor, ready to
intercept subsequent messages. Note that this means that any FF-A calls
involving the message buffers will fail until they are implemented in the
hypervisor, as passing them through to the SPM will no longer work.

For now this only supports a mailbox buffer size of 4 KiB, and assumes
that the stage-2 page table also uses 4 KiB pages.

Signed-off-by: Andrew Walbran <qwandor@google.com>