KVM: arm64: Provide {get,put}_page() stubs for early hyp allocator
In nVHE protected mode, the EL2 code uses a temporary allocator during
boot while re-creating its stage-1 page-table. Unfortunately, the
'hyp_vmmemap' is not ready to use at this stage and so refcounting pages
is not yet possible. Since the hypervisor stage-1 mappings are not
refcounted anyway (and are, in fact, never removed), this hasn't posted
a problem so far.

In preparation for allowing hypervisor stage-1 mappings to be removed,
provide stub implementations for {get,put}_page() in the early allocator.

No functional change.

Signed-off-by: Quentin Perret <qperret@google.com>
Signed-off-by: Will Deacon <will@kernel.org>