KVM: arm64: relay entropy requests from protected guests directly to secure
As pKVM does not trust the host, it should not be involved in the
handling of, or be able to observe the response to entropy requests
issues by protected guests.

Let's just assume that support for the SMC flavor of the ARM SMCCC TRNG
interface is a prerequisite for pKVM support, and pass the request on
directly to the secure firmware.

Signed-off-by: Ard Biesheuvel <ardb@google.com>