ARM: smp: initialize secondary CPU offset in startup code
In order to be able to use a per-CPU variable to carry the CPU number of
each core, initialize the per-CPU offset register from the secondary
startup code. This ensures that any reference to smp_processor_id() from
C code produces the correct value, even when called very early in the
startup sequence.

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>