selftests: Add test for sharing memory with TEE in secure world.
This uses FFA_MEM_SHARE to share a memory region with the secure world,
checking that the pKVM hypervisor (if present) processes it correctly.
For this to work there must be an SPM to talk to, which could be
provided by TF-A in EL3. An actual SP is not required.

Signed-off-by: Andrew Walbran <qwandor@google.com>