KVM: arm64: Add component ID for sharing with secure world.
When checking memory transitions, we need some way to identify the
secure world, alongside the host, hypervisor and guests. A single ID is
enough, as even if the system supports multiple secure partitions via
S-EL2, the hypervisor running in the normal world doesn't maintain their
page tables and so it isn't responsible for distinguishing them.

Signed-off-by: Andrew Walbran <qwandor@google.com>