KVM: arm64: Introduce SPMD lock and take it for memory transactions.
This lock is used to protect the EL3 FF-A mailbox buffers, to prevent
concurrent access to them for two different memory transactions. This is
not a big issue now when only the host is allowed to make FF-A calls,
but will become an issue in the future if guests are allowed to make
FF-A calls, so it's best to set things up to be safe.

Signed-off-by: Andrew Walbran <qwandor@google.com>