misc: dice: Add driver to forward secrets to userspace
Open Profile for DICE is a protocol for deriving unique secrets at boot
used by some Android devices. The firmware/bootloader hands over secrets
in a reserved memory region, which this driver takes ownership of and
exposes it to userspace via a misc device.

Userspace obtains the region's size using an ioctl and mmaps the memory
to its address space. This mapping cannot be write+shared, giving
userspace a guarantee that the secrets have not been overwritten by
another process.

Userspace can also issue an ioctl requesting that the memory be wiped by
the driver. Because both the kernel and userspace mappings use
write-combine semantics, all clients will observe the memory as zeroed
after the ioctl has returned.

Cc: Andrew Scull <ascull@google.com>
Cc: Will Deacon <will@kernel.org>
Signed-off-by: David Brazdil <dbrazdil@google.com>