selftests: Make fragmented memory share test work without pKVM too.
FF-A specifies that the sender ID for MEM_FRAG_TX must be specified at
the physical FF-A interface but must be 0 at the virtual FF-A interface.
This unfortunately means that EL1 needs to behave a little differently
depending on whether it is making FF-A calls to a hypervisor in EL2 or
an SPMD in EL3.

Signed-off-by: Andrew Walbran <qwandor@google.com>