dt-bindings: firmware: Add Open Profile for DICE
Add DeviceTree bindings for Open Profile for DICE, an open protocol for
measured boot. Firmware uses DICE to measure the hardware/software
combination and generates Compound Device Identifier (CDI) certificates.
These are stored in memory and the buffer is described in the DT as
a reserved memory region referenced by a compatible device node.

Signed-off-by: David Brazdil <dbrazdil@google.com>