KVM: arm64: Create EL2 shadow for VMs when pKVM is enabled
Regardless of whether a given VM is protected or unprotected, when pKVM
is enabled we must create an EL2 shadow for each VM so that the stage-2
page-tables are managed by the hypervisor instead of the host.

Create the EL2 shadow for a VM on the first-run of a vCPU when a shadow
is not already present.

Signed-off-by: Will Deacon <will@kernel.org>