ANDROID: KVM: arm64: pkvm: Add support for fragmented FF-A descriptors
FF-A memory descriptors may need to be sent in fragments when they don't
fit in the mailboxes. Doing so involves using the FRAG_TX and FRAG_RX
primitives defined in the FF-A protocol.

Add support in the pKVM FF-A relayer for fragmented descriptors by
monitoring outgoing FRAG_TX transactions and by buffering large
descriptors on the reclaim path.

Co-developed-by: Quentin Perret <qperret@google.com>
Signed-off-by: Andrew Walbran <qwandor@google.com>
Signed-off-by: Quentin Perret <qperret@google.com>
Bug: 171706629
Signed-off-by: Will Deacon <willdeacon@google.com>
[ qperret: fixed context conflicts and references to host_kvm.ffa and
  ffa_buffers that have been renamed ]
Signed-off-by: Quentin Perret <qperret@google.com>
Change-Id: I701f279cd4820abb0b6d7c2572ee28e0f943edad