KVM: arm64: Zap the PTE entries when the host reclaims ownership
When the host reclaims pages from the hypervisor in __pkvm_host_reclaim_page
the PTEs are updated to reflect the new owner. The newly installed PTE
has the value '0' and there is currently no memory associated with it.
When the refcount of a page table page falls below 1 we know that this
table has no children and we zap that entry.

Signed-off-by: Sebastian Ene <sebastianene@google.com>