KVM: arm64: Add nvhe_mem VM statistic
When using nVHE in protected mode, the host allocates memory for the
hypervisor to store shadow structures and the stage-2 page tables. Account
for those allocations in bytes, in a newly created VM statistic "nvhe_mem".

It is expected, on VM teardown to reclaim all that memory. Raise a warning
if not all the given memory is back.

Signed-off-by: Vincent Donnefort <vdonnefort@google.com>