jump_label: Add a generic interface to register a jump table
jump_label supports multiple jump tables. This intends originally to
support in-modules static keys (shared or not). This feature is interesting
for the arm64 nVHE hypervisor. The latter needs to both share keys with the
host and create its own.

Split the interface to allow any component, module or not, to register a
jump table. This newly introduced interface only expects a range where the
jump table can found and an set of optional callbacks.

Signed-off-by: Vincent Donnefort <vdonnefort@google.com>