KVM: arm64: Add clock support
EL2 has access to the same counter as EL1 is using for its sched_clock.
This makes it a easy way of generating clock values that are synchronized
with the host.

hyp_clock() returns the clock value in ns, similar to sched_clock().

hyp_clock_setup() will prepare the clock based on epoch values. If set from
the kernel, hyp_clock() and sched_clock() will be synchronized.

With untrusted epoch values from the host, hyp_clock() only intends to be
used for debugging and profiling purpose. And actually at the moment the
only planed user is tracing, hence a dependency on CONFIG_TRACING.

Signed-off-by: Vincent Donnefort <vdonnefort@google.com>